Add modular date parsing utilities with backward compatibility
